Move  code to a more standard location
