+++ tool/tctl/common/cmds.go
@@ -23,10 +23,10 @@ import (
 	"github.com/gravitational/teleport/tool/tctl/common/decision"
 	"github.com/gravitational/teleport/tool/tctl/common/loginrule"
 	"github.com/gravitational/teleport/tool/tctl/common/plugin"
+	"github.com/gravitational/teleport/tool/tctl/common/sso/configure"
+	"github.com/gravitational/teleport/tool/tctl/common/sso/tester"
 	"github.com/gravitational/teleport/tool/tctl/common/stableunixusers"
 	"github.com/gravitational/teleport/tool/tctl/common/top"
-	"github.com/gravitational/teleport/tool/tctl/sso/configure"
-	"github.com/gravitational/teleport/tool/tctl/sso/tester"
 )
 
 // Commands returns the set of available subcommands for tctl.
