+++ tool/tctl/common/sso/configure/command.go
