+++ tool/tctl/common/sso/configure/flags/attributes_to_roles.go
