+++ tool/tctl/common/sso/configure/oidc.go
@@ -35,8 +35,8 @@ import (
 	"github.com/gravitational/teleport/lib/auth/authclient"
 	"github.com/gravitational/teleport/lib/observability/otelhttp"
 	"github.com/gravitational/teleport/lib/utils"
-	"github.com/gravitational/teleport/tool/tctl/sso/configure/flags"
-	"github.com/gravitational/teleport/tool/tctl/sso/tester"
+	"github.com/gravitational/teleport/tool/tctl/common/sso/configure/flags"
+	"github.com/gravitational/teleport/tool/tctl/common/sso/tester"
 )
 
 type oidcPreset struct {
