+++ tool/tctl/common/sso/tester/command.go
