+++ tool/tctl/common/sso/tester/format.go
